Crate revmc_cranelift

revmc-cranelift

EVM bytecode compiler Cranelift backend.

This crate is currently not functional due to missing i256 support in Cranelift.

Re-exports

pub use cranelift;

pub use cranelift_jit;

pub use cranelift_module;

pub use cranelift_native;

Modules

pretty_clif 🔒

This module provides the CommentWriter which makes it possible to add comments to the written cranelift IR.

Structs

EvmCraneliftBackend

The Cranelift-based EVM bytecode compiler backend.

EvmCraneliftBuilder

The Cranelift-based EVM bytecode compiler function builder.

Symbols 🔒

Enums

ModuleWrapper 🔒

Functions

convert_intcc 🔒

convert_linkage 🔒

opt_level_flag 🔒