revmc_cranelift

Struct EvmCraneliftBackend

pub struct EvmCraneliftBackend {
    pub(crate) builder_context: FunctionBuilderContext,
    pub(crate) ctx: Context,
    pub(crate) module: ModuleWrapper,
    pub(crate) symbols: Symbols,
    pub(crate) opt_level: OptimizationLevel,
    pub(crate) comments: CommentWriter,
    pub(crate) functions: Vec<FuncId>,
}

The Cranelift-based EVM bytecode compiler backend.

Fields
builder_context: FunctionBuilderContext

The function builder context, which is reused across multiple FunctionBuilder instances.

ctx: Context

The main Cranelift context, which holds the state for codegen. Cranelift separates this from Module to allow for parallel compilation, with a context per thread, though this isn’t in the simple demo here.

module: ModuleWrapper

The module, with the jit backend, which manages the JIT’d functions.

symbols: Symbolsopt_level: OptimizationLevelcomments: CommentWriterfunctions: Vec<FuncId>
Implementations

impl EvmCraneliftBackend

pub fn is_supported() -> Result<(), &'static str>

Returns Ok(()) if the current architecture is supported, or Err(()) if the host machine is not supported in the current configuration.

pub fn new(aot: bool, opt_level: OptimizationLevel) -> Self

Creates a new instance of the JIT compiler.

Panics

Panics if the current architecture is not supported. See is_supported.

pub(crate) fn finish_module(&mut self) -> Result<Option<ObjectModule>>